+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- setup tools subpackage
+- validation_tools module created
+- Defined is_valid_slot function for dynamic __slots__ creation.
+- Defined canonical_slots function for dynamic __slots__ creation.
+
 ### Changed
 - Update ruff and sync uv.lock