+++ CHANGELOG.md
@@ -14,8 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added the protocols module with Freeze and Freezable protocols.
 
 ### Changed
-- Update ruff and sync uv.lock
+- Update ruff v0.15.0 -> v0.15.1 & uv.lock
 - Update librt v0.8.0 -> v0.8.1 & uv.lock
+
+### Fixed
+- minor doc fixes
 
 ## [0.1.2] - 2026-02-12
 ### Fixed