+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Update ruff v0.15.0 -> v0.15.1 & uv.lock
 - Update librt v0.8.0 -> v0.8.1 & uv.lock
+- Update ruff v0.15.1 -> v0.15.2 & uv.lock
 
 ### Fixed
 - minor doc fixes