+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update protocols module with the Immutable protocol
 - Update taph.__init__ module with new names
 - Update ruff v0.15.2 -> v0.15.04
+- Update pyproject.tool.ruff.lint with more coverage and sorted list
+- Update taph.protocols doc strings
 
 ### Fixed
 - minor doc fixes